+++ internal/api/handlers/ws.go
@@ -8,23 +8,13 @@ type WSHandler struct {
 	*BaseHandler
 }
 
-func (h *WSHandler) HandleConsole(w http.ResponseWriter, r *http.Request) {
+func (h *WSHandler) HandleWebSocket(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 	if id == "" {
 		http.Error(w, "Missing ID", http.StatusBadRequest)
 		return
 	}
 
-	hub := h.HubManager.GetHub(id)
-	hub.ServeWs(w, r)
-}
-
-func (h *WSHandler) HandleProgress(w http.ResponseWriter, r *http.Request) {
-	id := r.PathValue("id")
-	if id == "" {
-		http.Error(w, "Missing ID", http.StatusBadRequest)
-		return
-	}
 	hub := h.HubManager.GetHub(id)
 	hub.ServeWs(w, r)
 }